+++ src/components/navigation-bar/navigation-bar.style.ts
@@ -1,8 +1,19 @@
 import styled, { css } from "styled-components";
-import { padding, flexbox } from "styled-system";
+import { padding, flexbox, PaddingProps, FlexboxProps } from "styled-system";
 import { baseTheme } from "../../style/themes";
+import { StickyPosition, NavigationType } from "./navigation-bar.component";
 
-const StyledNavigationBar = styled.nav`
+type StyledNavigationBarProps = PaddingProps &
+  FlexboxProps & {
+    /** Color scheme of navigation component */
+    navigationType?: NavigationType;
+    /** Defines the position of sticky navigation bar */
+    stickyPosition?: StickyPosition;
+    /** Defines the offset of sticky navigation bar */
+    stickyOffset?: string;
+  };
+
+const StyledNavigationBar = styled.nav<StyledNavigationBarProps>`
   display: flex;
   align-items: center;
   padding: 0 40px;
@@ -45,14 +56,30 @@ const StyledNavigationBar = styled.nav`
 
   ${({ navigationType, theme }) => css`
     min-height: 40px;
-    background-color: ${theme.navigationBar.light.background};
-    border-bottom: 1px solid ${theme.navigationBar.light.borderBottom};
+
+    ${navigationType === "light" &&
+    css`
+      background-color: ${theme.navigationBar.light.background};
+      border-bottom: 1px solid ${theme.navigationBar.light.borderBottom};
+    `}
 
     ${navigationType === "dark" &&
     css`
       background-color: ${theme.navigationBar.dark.background};
       border-bottom: 1px solid ${theme.navigationBar.dark.borderBottom};
       color: ${theme.colors.white};
+    `}
+
+    ${navigationType === "black" &&
+    css`
+      background-color: ${theme.navigationBar.black.background};
+      color: ${theme.colors.white};
+    `}
+
+    ${navigationType === "white" &&
+    css`
+      background-color: ${theme.colors.white};
+      border-bottom: 1px solid ${theme.navigationBar.white.borderBottom};
     `}
   `}
 `;