+++ src/components/progress-tracker/progress-tracker.d.ts
@@ -24,11 +24,6 @@ export interface ProgressBarProps extends MarginProps {
   currentProgressLabel?: string;
   /** Value to display as the maximum progress limit. */
   maxProgressLabel?: string;
-  /**
-   * Sets the colour of the bar that shows the current progress.
-   * The "traffic" variant changes the colour of status bar depending on current progress.
-   * */
-  variant?: "default" | "traffic";
   /** The orientation of the component. */
   orientation?: "horizontal" | "vertical";
   /** The direction the bar should move as progress increases, only applies in vertical orientation. */