+++ src/components/select/option/option.spec.js
@@ -2,7 +2,6 @@ import React from "react";
 import TestRenderer from "react-test-renderer";
 import { shallow, mount } from "enzyme";
 import Option from "./option.component";
-import { baseTheme } from "../../../style/themes";
 
 describe("Option", () => {
   it("renders properly", () => {
@@ -22,7 +21,7 @@ describe("Option", () => {
       const props = { value: "1", text: "foo", isHighlighted: true };
       expect(renderOption(props, mount)).toHaveStyleRule(
         "background-color",
-        baseTheme.select.selected
+        "var(--colorsUtilityMajor200)"
       );
     });
   });
@@ -39,7 +38,7 @@ describe("Option", () => {
       const props = { value: "1", text: "foo" };
       expect(renderOption(props, mount)).toHaveStyleRule(
         "background-color",
-        baseTheme.select.selected,
+        "var(--colorsUtilityMajor200)",
         { modifier: ":hover" }
       );
     });