+++ src/components/tooltip/tooltip-pointer.style.js
@@ -5,7 +5,9 @@ import { toColor } from "../../style/utils/color";
 
 const pointerColor = (type, theme, bgColor) => {
   if (bgColor) return toColor(theme, bgColor);
-  return type === "error" ? theme.colors.error : theme.colors.black;
+  return type === "error"
+    ? "var(--colorsSemanticNegative500)"
+    : "var(--colorsSemanticNeutral500)";
 };
 const StyledTooltipPointer = styled.div`
   ${({ position, theme, type, bgColor }) => css`
@@ -18,7 +20,7 @@ const StyledTooltipPointer = styled.div`
       border-left: 8px solid transparent;
       border-right: 8px solid transparent;
       border-top: 8px solid ${pointerColor(type, theme, bgColor)};
-      bottom: -${theme.spacing}px;
+      bottom: calc(-1 * var(--spacing100));
       @-moz-document url-prefix() {
         bottom: -7px;
       }
@@ -29,7 +31,7 @@ const StyledTooltipPointer = styled.div`
       border-left: 8px solid transparent;
       border-right: 8px solid transparent;
       border-bottom: 8px solid ${pointerColor(type, theme, bgColor)};
-      top: -${theme.spacing}px;
+      top: calc(-1 * var(--spacing100));
       @-moz-document url-prefix() {
         top: -7px;
       }
@@ -40,7 +42,7 @@ const StyledTooltipPointer = styled.div`
       border-top: 8px solid transparent;
       border-bottom: 8px solid transparent;
       border-right: 8px solid ${pointerColor(type, theme, bgColor)};
-      left: -${theme.spacing}px;
+      left: calc(-1 * var(--spacing100));
       @-moz-document url-prefix() {
         left: -7px;
       }
@@ -51,7 +53,7 @@ const StyledTooltipPointer = styled.div`
       border-top: 8px solid transparent;
       border-bottom: 8px solid transparent;
       border-left: 8px solid ${pointerColor(type, theme, bgColor)};
-      right: -${theme.spacing}px;
+      right: calc(-1 * var(--spacing100));
       @-moz-document url-prefix() {
         right: -7px;
       }
@@ -68,6 +70,7 @@ StyledTooltipPointer.propTypes = {
   position: PropTypes.oneOf(["bottom", "left", "right", "top"]),
   theme: PropTypes.object,
   type: PropTypes.string,
+  bgColor: PropTypes.string,
 };
 
 export default StyledTooltipPointer;